requirements.txt
torch>=2.0.0
numpy>=1.23.0
pyyaml>=6.0